znovu dao 1

+++ tests/Integration/Event/Model/Dao/EventLinkDaoTest.php
@@ -29,11 +29,11 @@ class EventLinkDaoTest extends AppRunner {
      */
     private $dao;
 
-    private static $eventLink1PrimaryKey;
-    private static $eventLink2PrimaryKey;
+    private static $eventLink1PrimaryKeyTouple;
+    private static $eventLink2PrimaryKeyTouple;
 
-    private static $eventPrimaryKey;
-    private static $eventLinkPhasePrimaryKey;
+    private static $eventPrimaryKeyTouple;
+    private static $eventLinkPhasePrimaryKeyTouple;
 
 
     public static function setUpBeforeClass(): void {
@@ -48,7 +48,7 @@ class EventLinkDaoTest extends AppRunner {
         $rowData = new RowData();
         $rowData->import( ['text' => 'Konečná fáze'   ] );
         $eventLinkPhaseDao->insert($rowData);
-        self::$eventLinkPhasePrimaryKey = $eventLinkPhaseDao->getLastInsertedPrimaryKey();
+        self::$eventLinkPhasePrimaryKeyTouple = $eventLinkPhaseDao->getLastInsertedPrimaryKey();
     }
 
 
@@ -71,25 +71,35 @@ class EventLinkDaoTest extends AppRunner {
                 (new TestDbEventsContainerConfigurator())->configure(new Container())
             );
 
-
+        //smazat -- id ----
+        /** @var EventLinkDao $eventLinkDao */
+        $eventLinkDao = $container->get(EventLinkDao::class);
+        $eventRow = $eventLinkDao->get(self::$eventLink1PrimaryKeyTouple);  //PS nově mažu oba linky v testDelete
+        if ($eventRow) {
+            $eventLinkDao->delete($eventRow);
+        }       
         //smazat -- id2 ----
         /** @var EventLinkDao $eventLinkDao */
         $eventLinkDao = $container->get(EventLinkDao::class);
-        $eventRow = $eventLinkDao->get(self::$eventLink2PrimaryKey);  //PS nově mažu oba linky v testDelete
-        $eventLinkDao->delete($eventRow);
-
+        $eventRow = $eventLinkDao->get(self::$eventLink2PrimaryKeyTouple);  //PS nově mažu oba linky v testDelete
+        if ($eventRow) {
+            $eventLinkDao->delete($eventRow);
+        }
         //smazat -- event ----
         /** @var EventDao $eventDao */
         $eventDao = $container->get(EventDao::class);
-        $eventRow = $eventDao->get(self::$eventPrimaryKey);
-        $eventDao->delete($eventRow);
-
+        $eventRow = $eventDao->get(self::$eventPrimaryKeyTouple);
+        if ($eventRow) {
+            $eventDao->delete($eventRow);
+        }
+        
         //smazat -- event_link_phase ----
          /** @var EventLinkPhaseDao $eventLinkPhaseDao */
         $eventLinkPhaseDao = $container->get(EventLinkPhaseDao::class);
-        $eventLinkPhaseRow = $eventLinkPhaseDao->get(self::$eventLinkPhasePrimaryKey);
-        $eventLinkPhaseDao->delete($eventLinkPhaseRow);
-
+        $eventLinkPhaseRow = $eventLinkPhaseDao->get(self::$eventLinkPhasePrimaryKeyTouple);
+        if ($eventLinkPhaseRow) {
+         $eventLinkPhaseDao->delete($eventLinkPhaseRow);
+        }
     }
 
 
@@ -101,23 +111,30 @@ class EventLinkDaoTest extends AppRunner {
      public function testInsert() {
         //-----------------id1-------
         $rowData = new RowData();
-        $rowData->offsetSet('show', 1);
-        $rowData->offsetSet('href', "EventLinkDaoTesthttpassdrooosaasdas_1");
-        $rowData->offsetSet('link_phase_id_fk', self::$eventLinkPhasePrimaryKey ['id'] ) ;
+//        $rowData->offsetSet('show', 1);            //slo by zde taky takto nastavit
+//        $rowData->offsetSet('href', "EventLinkDaoTesthttpassdrooosaasdas_1");
+//        $rowData->offsetSet('link_phase_id_fk', self::$eventLinkPhasePrimaryKey ['id'] ) ;
+        $rowData->import( [ 'show' => 1, 
+                            'href' => "EventLinkDaoTesthttpassdrooosaasdas_1", 
+                            'link_phase_id_fk' => self::$eventLinkPhasePrimaryKeyTouple ['id'] ] );        
         $this->dao->insert($rowData);
-        self::$eventLink1PrimaryKey =  $this->dao->getLastInsertedPrimaryKey();
-        $this->assertIsArray(self::$eventLink1PrimaryKey);
+        self::$eventLink1PrimaryKeyTouple =  $this->dao->getLastInsertedPrimaryKey();
+        $this->assertIsArray(self::$eventLink1PrimaryKeyTouple);
         $numRows = $this->dao->getRowCount();
         $this->assertEquals(1, $numRows);
 
         //-----------------id2-------
         $rowData = new RowData();
-        $rowData->offsetSet('show', 2);
-        $rowData->offsetSet('href', "EventLinkDaoTesthttpassdrooo_2");
-        $rowData->offsetSet('link_phase_id_fk', self::$eventLinkPhasePrimaryKey ['id'] );
+//        $rowData->offsetSet('show', 2);            //slo by zde taky takto nastavit
+//        $rowData->offsetSet('href', "EventLinkDaoTesthttpassdrooo_2");
+//        $rowData->offsetSet('link_phase_id_fk', self::$eventLinkPhasePrimaryKey ['id'] );
+        $rowData->import( [ 'show' => 2, 
+                            'href' => "EventLinkDaoTesthttpassdrooo_2", 
+                            'link_phase_id_fk' => self::$eventLinkPhasePrimaryKeyTouple ['id'] ] );        
+
         $this->dao->insert($rowData);
-        self::$eventLink2PrimaryKey =  $this->dao->getLastInsertedPrimaryKey();
-        $this->assertIsArray(self::$eventLink2PrimaryKey);
+        self::$eventLink2PrimaryKeyTouple =  $this->dao->getLastInsertedPrimaryKey();
+        $this->assertIsArray(self::$eventLink2PrimaryKeyTouple);
         $numRows = $this->dao->getRowCount();
         $this->assertEquals(1, $numRows);
 
@@ -126,31 +143,38 @@ class EventLinkDaoTest extends AppRunner {
         /** @var EventDao $eventDao */
         $eventDao = $this->container->get( EventDao::class );
         $rowData = new RowData();
-        $rowData->offsetSet('published', 1);
-        $rowData->offsetSet('start', "2011-01-01 15:03:01" );
-        $rowData->offsetSet('end', "2011-01-02 1:00:00");
-        $rowData->offsetSet('enroll_link_id_fk', self::$eventLink1PrimaryKey['id'] );
-        $rowData->offsetSet('enter_link_id_fk', self::$eventLink2PrimaryKey['id'] );
-        $rowData->offsetSet('event_content_id_fk', null);
+//        $rowData->offsetSet('published' , 1);                         //slo by zde taky takto nastavit
+//        $rowData->offsetSet('start', "2011-01-01 15:03:01" );
+//        $rowData->offsetSet('end', "2011-01-02 1:00:00");
+//        $rowData->offsetSet('enroll_link_id_fk', self::$eventLink1PrimaryKey['id'] );
+//        $rowData->offsetSet('enter_link_id_fk', self::$eventLink2PrimaryKey['id'] );
+//        $rowData->offsetSet('event_content_id_fk', null);
+        $rowData->import( [ 'published' => 1, 
+                            'start' => "2011-01-01 15:03:01", 
+                            'end' => "2011-01-02 1:00:00",
+                            'enroll_link_id_fk' => self::$eventLink1PrimaryKeyTouple['id'],
+                            'enter_link_id_fk' => self::$eventLink2PrimaryKeyTouple['id'],
+                            'event_content_id_fk' => null  ] );        
+        
         $eventDao->insert($rowData);
-        self::$eventPrimaryKey =  $eventDao->getLastInsertedPrimaryKey();
+        self::$eventPrimaryKeyTouple =  $eventDao->getLastInsertedPrimaryKey();
     }
 
 
     public function testGetExistingRow() {
-        $eventRow = $this->dao->get(self::$eventLink1PrimaryKey);
+        $eventRow = $this->dao->get(self::$eventLink1PrimaryKeyTouple);
         $this->assertInstanceOf(RowDataInterface::class, $eventRow);
     }
 
     public function test4Columns() {
-        $eventRow = $this->dao->get(self::$eventLink1PrimaryKey);
+        $eventRow = $this->dao->get(self::$eventLink1PrimaryKeyTouple);
         $this->assertCount(4, $eventRow);
     }
 
 
 
     public function testUpdate() {
-        $eventLinkRow = $this->dao->get(self::$eventLink1PrimaryKey);
+        $eventLinkRow = $this->dao->get(self::$eventLink1PrimaryKeyTouple);
         $this->assertIsString( $eventLinkRow['href']);
         $ret = $eventLinkRow['href'];
         //
@@ -161,7 +185,7 @@ class EventLinkDaoTest extends AppRunner {
         $this->assertEquals(1, $this->dao->getRowCount());
 
         $this->setUp();
-        $eventRowRereaded = $this->dao->get(self::$eventLink1PrimaryKey);
+        $eventRowRereaded = $this->dao->get(self::$eventLink1PrimaryKeyTouple);
         $this->assertEquals($eventLinkRow, $eventRowRereaded);
         $this->assertStringContainsString('nahrada-ooo-', $eventRowRereaded['href']);
     }
@@ -174,31 +198,32 @@ class EventLinkDaoTest extends AppRunner {
     }
 
     public function testDelete() {
-        $eventRow = $this->dao->get(self::$eventLink1PrimaryKey);
+        $eventRow = $this->dao->get(self::$eventLink1PrimaryKeyTouple);
         $this->dao->delete($eventRow);
         $this->assertEquals(1, $this->dao->getRowCount());
-        $eventRow = $this->dao->get(self::$eventLink2PrimaryKey);
+        $eventRow = $this->dao->get(self::$eventLink2PrimaryKeyTouple);
         $this->dao->delete($eventRow);
         $this->assertEquals(1, $this->dao->getRowCount());
 
         $this->setUp();
 
-        $eventRow = $this->dao->get(self::$eventLink1PrimaryKey);
+        $eventRow = $this->dao->get(self::$eventLink1PrimaryKeyTouple);
         $this->assertNull($eventRow);
-        $eventRow = $this->dao->get(self::$eventLink2PrimaryKey);
+        $eventRow = $this->dao->get(self::$eventLink2PrimaryKeyTouple);
         $this->assertNull($eventRow);
 
         // kontrola SET
         // kontrola, že  se deletem v event_link tabulce se  nastavilo v tabulce event  event.x_link_id_fk = null
         /** @var EventDao $eventDao */
         $eventDao = $this->container->get( EventDao::class );
-        $eventData =  $eventDao->get (self::$eventPrimaryKey);
+        $eventData =  $eventDao->get (self::$eventPrimaryKeyTouple);
         $this->assertNull( $eventData['enroll_link_id_fk'] );
         $this->assertNull( $eventData['enter_link_id_fk'] );
 
 
         //---------------------------------------
-        //smazat v event  id2 vetu - provede tearDownAfterClass
+        //smazat v event  vetu - provede tearDownAfterClass
+        //smazat v eventlink 2.vetu - provede tearDownAfterClass
     }
